Feature ot29845

+++ src/main/java/com/alkemy/ong/service/impl/NewsServiceImpl.java
@@ -44,9 +44,14 @@ public class NewsServiceImpl implements INewsService {
         return newsDto;
     }
 
+    @Override
     public void deleteById(Long id){
+        if(!findById(id).isPresent()){
+            throw new ResourceNotFoundException(message.getMessage("news.notFound", null, Locale.US));
+        }
         repo.deleteById(id);
     }
+
     public boolean existsById(Long id){
         return repo.existsById(id);
     }
